Moving PekException back to the service layer
It is exclusively used from the service layer, there is no
point in putting it in the util package.
