+++ sch-pek-ejb-impl/src/main/java/hu/sch/ejb/AccountManagerBean.java
@@ -13,8 +13,8 @@ import hu.sch.services.SystemManagerLocal;
 import hu.sch.services.UserManagerLocal;
 import hu.sch.util.config.Configuration;
 import hu.sch.services.exceptions.DuplicatedUserException;
-import hu.sch.util.exceptions.PekErrorCode;
-import hu.sch.util.exceptions.PekException;
+import hu.sch.services.exceptions.PekErrorCode;
+import hu.sch.services.exceptions.PekException;
 import hu.sch.util.hash.Hashing;
 import java.io.UnsupportedEncodingException;
 import java.nio.charset.StandardCharsets;
