+++ sch-pek-ejb-impl/src/main/java/hu/sch/ejb/ValuationManagerBean.java
@@ -23,8 +23,8 @@ import hu.sch.services.SystemManagerLocal;
 import hu.sch.services.UserManagerLocal;
 import hu.sch.services.ValuationManagerLocal;
 import hu.sch.services.exceptions.NoSuchAttributeException;
-import hu.sch.util.exceptions.PekException;
-import hu.sch.util.exceptions.PekErrorCode;
+import hu.sch.services.exceptions.PekException;
+import hu.sch.services.exceptions.PekErrorCode;
 import hu.sch.services.exceptions.UserNotFoundException;
 import hu.sch.services.exceptions.valuation.AlreadyModifiedException;
 import hu.sch.services.exceptions.valuation.NoExplanationException;
