+++ sch-pek-ejb-impl/src/main/java/hu/sch/ejb/image/ImageProcessor.java
@@ -3,8 +3,8 @@ package hu.sch.ejb.image;
 import hu.sch.util.config.ImageUploadConfig;
 import hu.sch.domain.user.ProfileImage;
 import hu.sch.domain.user.User;
-import hu.sch.util.exceptions.PekException;
-import hu.sch.util.exceptions.PekErrorCode;
+import hu.sch.services.exceptions.PekException;
+import hu.sch.services.exceptions.PekErrorCode;
 import hu.sch.util.hash.Hashing;
 import hu.sch.util.net.MediaType;
 import java.io.File;
