+++ sch-pek-ejb-services/src/main/java/hu/sch/services/exceptions/PekErrorType.java
@@ -1,4 +1,4 @@
-package hu.sch.util.exceptions;
+package hu.sch.services.exceptions;
 
 /**
  * Type for PekErrorCode. Every error code has a 2-character type.
