+++ sch-pek-internal-api/src/main/java/hu/sch/api/providers/DefaultExceptionHandler.java
@@ -1,7 +1,7 @@
 package hu.sch.api.providers;
 
 import hu.sch.api.response.PekError;
-import hu.sch.util.exceptions.PekException;
+import hu.sch.services.exceptions.PekException;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.ext.ExceptionMapper;
