+++ sch-pek-internal-api/src/main/java/hu/sch/api/providers/NotFoundExceptionHandler.java
@@ -1,7 +1,7 @@
 package hu.sch.api.providers;
 
 import hu.sch.api.response.PekError;
-import hu.sch.util.exceptions.PekErrorCode;
+import hu.sch.services.exceptions.PekErrorCode;
 import javax.ws.rs.NotFoundException;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
