+++ sch-pek-internal-api/src/main/java/hu/sch/api/user/UsersAvatar.java
@@ -9,8 +9,8 @@ import hu.sch.api.response.PekSuccess;
 import hu.sch.domain.user.ProfileImage;
 import hu.sch.domain.user.User;
 import hu.sch.util.config.Configuration;
-import hu.sch.util.exceptions.PekErrorCode;
-import hu.sch.util.exceptions.PekException;
+import hu.sch.services.exceptions.PekErrorCode;
+import hu.sch.services.exceptions.PekException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.logging.Level;
